samples/agent/adk/real_estate_agent/pyproject.toml
[project]
name = "a2ui-real-estate-agent"
version = "0.1.0"
description = "Real Estate Agent with Google Search grounding and Google Maps integration."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "google-adk>=1.9.0",
    "google-genai>=1.27.0",
    "googlemaps>=4.10.0",
    "python-dotenv>=1.1.0",
    "a2ui",
    "fastapi",
    "uvicorn",
    "jsonschema",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
